refactor(geo): move package to baseapp.geo namespace
Mirror baseapp.files (PR #442): module path baseapp.geo, app label stays
baseapp_geo so swapper settings, backend slot keys and consumer
migrations are unchanged. Entry point now baseapp_geo = "baseapp.geo.plugin:GeoPlugin";
AppConfig aligned with namespace convention (BigAutoField).
