+++ pyproject.toml
@@ -213,7 +213,7 @@ baseapp_mentions = "baseapp_mentions.plugin:MentionsPlugin"
 baseapp_follows = "baseapp_follows.plugin:FollowsPlugin"
 baseapp_reports = "baseapp_reports.plugin:ReportsPlugin"
 baseapp_ratings = "baseapp_ratings.plugin:RatingsPlugin"
-baseapp_geo = "baseapp_geo.plugin:GeoPlugin"
+baseapp_geo = "baseapp.geo.plugin:GeoPlugin"
 baseapp_reactions = "baseapp_reactions.plugin:ReactionsPlugin"
 baseapp_blocks = "baseapp_blocks.plugin:BlocksPlugin"
 baseapp_content_feed = "baseapp.content_feed.plugin:ContentFeedPlugin"
